feat: make compatible with foundry V9
Additionally, this introduces a new setting to set the required role to be able to create
effects.

BREAKING CHANGE: In foundry V9, it's not possible anymore to manipulate
the permissions in `CONST`. For that reason, it was necessary to switch
to using a setting instead. Unfortunately, it is not easily possible to
to migrate from the old way to the new way, so users will have to adapt
their settings to match what they had configured previously.