fix: load special effect files lazily
With this change, special effect files are only loaded when the preview box is hovered or when the
effect is actually played. Nothing is loaded automatically, not even the metadata. This should help
to improve the experience when using lot's of special effects such as when using the effects from
JB2A.

Closes #209